Updated styles.css and index.html
I removed the br tags from index.html, and I added a width and height to the div that contains the navbar. This way, the h1 title and the navbar won't mix with each other, and I won't have to use the br tags in index.html as a workaround to avoid the navbar and the h1 title to mix with each other. I also added a media query so that the image looks well on screens with 1536px of width. Finally, I added some top padding to the h1 title.