Updated styles.css and index.html
I centered vertically and horizontally the h1 title of the landing page, and it's responsive (it stays at the center of the screen in all screen sizes). The navbar and the portfolio button remain at the top and the bottom of the screen. I just need to make the portfolio button and the navbar elements responsive so that the distance that they are from the top and bottom margins of the screen are responsive.