Updated styles.css
I vertically centered the small links of the navbar, and it's responsive. I need to now vertically center the Nova Logo (which can be done by making the height of the logo the same height of the navbar div). Also, I will create a variable for the logo height, the navbar div height, and the line-height property so that all navbar elements are vertically centered.