Created Side Navbar (BUGGY)
I am able to toggle a side navbar in phone screens. However, if I press the 'x' or 'open' buttons in large screens, the small navbar links will look messy on the screen. Also, I am using an ID for 'mySidenavbar', when in reality I need to use it as a class.