Fixed bug in script.js
I was able to fix the bug that caused the navbar text to get messed up in larger screens if the user decided to touch the hamburguer menu button. However, the bug is only fixed if the user decides to close the menu on a phone screen before switching into a larger screen. If the user decides to leave the hamburguer menu open before switching into a larger screen, the bug will still be present.