Updated styles and index.html
I copied and pasted the 4 navbar links into the index.html file to create a copy of the small navbar links. I did this since the 'li' tags from the navbar were giving me trouble whenever I tried to change the styles of the navbar links while they were inside of the side navbar. Now, the side navbar elements are in the appropriate order, and they are shown vertically.