Added Images and Edited Styles
I added the two images of Nova's office, and I fixed the body section with the white background so that it uses flexbox instead of vw to be divided into 2 halves. Now the body is properly displaying me text in the left and an image on the right in computer screens, and the txt on top of the image in tablet and phone screens.