Created Footer (BUGGY)
I made a placeholder footer. No white space will appear below the footer in any device. There's a bug in which the hand icon appears (as if the link is clickable) to the sides of the 'back to top' link, when it should only appear while hovering the 'back to top' text.