Bug Fix in Home
I modified the code of the 'Check our portfolio' button so that the user can click anywhere in the button and still reach the portfolio page. Before that, the button didn't work if I clicked outside of the text of the button. I also added a link to the home page on the Nova logo in the portfolio.html page.