Updated Side Margins
I updated the side margins of the 'Back to top' links on all pages. It also updated the side margins of some versions of the footer. I need to fix a bug in which the 'Back to top' link and the 'Portfolio' link on the footer have different margins in iPads in horzontal mode.